chore: initial commit for implementing routes (user)
