+++ frontend/src/App.css
@@ -127,6 +127,8 @@ html {
 .common_input {
   height: 1.4vw;
   font-size: 1.25vw;
+  width: 95%;
+  margin-bottom: 5px;
 }
 
 .login_form div {
@@ -199,6 +201,18 @@ input[type="checkbox"] {
   border-bottom-left-radius: 10px;
   border-bottom-right-radius: 10px;
   box-shadow: 0px 0px 10px 6px rgba(0, 0, 0, 0.2);
+}
+
+.select a {
+  display: flex;
+  padding-inline: 10px;
+  text-decoration: none;
+  color: white;
+  flex: 1 1;
+  align-items: center;
+  flex-wrap: wrap;
+  flex-direction: column;
+  justify-content: center;
 }
 
 .select h3 {
