+++ frontend/src/components/menu.js
@@ -1,23 +1,27 @@
 import { useEffect, useState } from "react";
+import { NavLink, Outlet, useNavigate } from "react-router-dom";
+
 
 export default function Menus(props){
   const menus = props.list;
   
   const [selectedMenu, setSelectedMenu] = useState(null);
-
+  // const navigate = useNavigate();
   // sets the default menu to home
   useEffect(() => {
     const homeMenu = menus.find(menu => menu.name.toLowerCase() === 'home');
     if (homeMenu) {
         setSelectedMenu(homeMenu.id);
+        // navigate('user/home')
         props.handleSelectClick(homeMenu.id);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []); // empty array means this useEffect will only run once
 
-  const handleMenuClick = (menu_id) => {
-    setSelectedMenu(menu_id);
-    props.handleSelectClick(menu_id); // Notify the parent component
+  const handleMenuClick = (menu) => {
+    setSelectedMenu(menu.id);
+    // navigate(`${menu.url}`);
+    props.handleSelectClick(menu.id); // Notify the parent component
   }
 
   return (
@@ -29,8 +33,17 @@ export default function Menus(props){
         </div>
         <div className="select">
           {menus.map(menu => (
+            // <NavLink 
+            //   to={menu.url} 
+            //   key={menu.id} 
+            //   id={menu.name.toLowerCase}
+            //   onClick={() => handleMenuClick(menu)}
+            //   className={selectedMenu === menu.id ? 'selected_menu' : ''}
+            // >
+            //   {menu.name}
+            // </NavLink>
             <h3 key={menu.id} id={menu.name.toLowerCase()} 
-            onClick={() => handleMenuClick(menu.id)}
+            onClick={() => handleMenuClick(menu)}
             className={selectedMenu === menu.id ? 'selected_menu' : ''}> 
               {menu.name}
             </h3>
